+++ R/fcwt.R
@@ -63,7 +63,9 @@
 #'
 #' @param n_threads
 #'  Number of threads used by the computation, if supported by your platform.
-#'  Defaults to 2 threads (to accomodate CRAN requirements).
+#'  Defaults to 2 threads (to accommodate CRAN requirements).
+#'  If [openmp_enabled()] returns `FALSE`, this argument is ignored, and
+#'  only a single thread is used.
 #'
 #' @return
 #'  The spectogram, a numeric real-valued matrix with dimensions