+++ tests/testthat/test-openmp.R
@@ -1,0 +1,5 @@
+test_that("openmp() works", {
+  expect_true(
+    openmp_enabled() == TRUE |  openmp_enabled() == FALSE
+  )
+})